refactor: 불필요한 data 전달하는 DTO 정리
