+++ src/main/java/com/example/aitalk/domain/member/dto/MemberJoinRequestDTO.java
@@ -1,4 +1,4 @@
-package com.example.aitalk.domain.member.dto.login;
+package com.example.aitalk.domain.member.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -6,12 +6,13 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import org.springframework.web.multipart.MultipartFile;
 
 @Data
 @Getter
 @Setter
 @NoArgsConstructor
-public class MemberLoginRequestDTO {
+public class MemberJoinRequestDTO {
 
     // 유효성 검사
     @NotBlank(message = "이메일은 필수 입력 값입니다.")
@@ -20,5 +21,9 @@ public class MemberLoginRequestDTO {
 
     @NotBlank(message = "비밀번호는 필수 입력 값입니다.")
     private String password; // 비밀번호 (VARCHAR(255))
+
+    private String nickname; // 닉네임 (NULL 허용, VARCHAR(255))
+
+    private MultipartFile profileImage;  // 이미지 파일 필드 추가
 
 }
