--- src/main/java/com/example/aitalk/domain/member/dto/login/MemberLoginResponseDTO.java
@@ -1,20 +1,0 @@
-package com.example.aitalk.domain.member.dto.login;
-
-import lombok.Builder;
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-@Data
-public class MemberLoginResponseDTO {
-
-    int statusCode; // 응답 상태 코드
-
-    String email; // 가입한 사용자 이메일
-
-    String message; // 응답 메시지
-    Long userId;
-}
